Turn on warnings as errors for our compile targets only (#23)
* Fix up backend to have all warnings turned on

* exclude for backend

* fix the public api test

* fix all frontend warnings

* fix formatting

* no point in separate link targets

* set to maybe_unused instead of void cast

* use unused for node

* add 2 more

[ROCm/hipDNN commit: 10671a72aa6aa4541b2bd88f43099422f483bd15]

+++ frontend/include/hipdnn_frontend/utilities.hpp
@@ -24,7 +24,7 @@ inline error_t find_common_shape(const std::vector<std::vector<int64_t>>& input_
         return {error_code_t::INVALID_VALUE, "Input shapes cannot be empty"};
     }
 
-    int64_t dims
+    size_t dims
         = std::max_element(input_shapes.begin(),
                            input_shapes.end(),
                            [](const std::vector<int64_t>& a, const std::vector<int64_t>& b) {
